Renamed to HuffmanCoding
